user 表结构修改

+++ web/src/main/resources/sql/19-02-12_auth.sql
@@ -5,13 +5,13 @@ create table my_user
   password varchar(20) default '' not null comment '密码',
   nick_name varchar(20) default '' not null comment '昵称',
   salt_value varchar(20) default '' not null comment '加密盐值',
-  status int(11) default 0 not null comment '0:正常 1:冻结',
+  status tinyint(2) default 0 not null comment '0:正常 1:冻结',
   remark varchar(200) default '' not null comment '备注',
   add_uid varchar(20) default '' not null comment '添加人',
   add_time timestamp default CURRENT_TIMESTAMP not null comment '注册时间',
   update_uid varchar(20) default '' not null comment '修改人',
   update_time timestamp default CURRENT_TIMESTAMP not null comment '修改时间',
-  is_delete int(11) default 0 not null comment '0: 删除 1: 未删除',
+  is_delete tinyint(2) default 0 not null comment '0: 删除 1: 未删除',
   last_update_time timestamp default CURRENT_TIMESTAMP not null comment '最后更新时间',
   constraint my_user_pk
     primary key (id)
